Rename `main` submodule to `core`
